fix: Build an i386 Linux executable using CMake
Problem: First, the CMake file `cmake/FindSDL2_ttf.cmake` was not able
to find the include path and the architecture defaulted to `x86_64`.
Second, some paths were assumed to be valid Windows paths. Third, the
`Dockerfile` is out of date and did not include the necessary
dependencies to build. And lastly, there was a missing definition of
`dvl::net::udp_p2p::poll()`.

Solution: First, `cmake/FindSDL2_ttf.cmake` was looking for
`SDL_mixer.h`. It should be looking for `SDL_ttf.h`. Hard code the
architecture and processor to i386. Add an empty
`dvl::net::udp_p2p::poll()` function.

Fix the computed paths to use `/` and not `\` to delimit paths.

And lastly, add the necessary dependencies in the `Dockerfile`.

Testing: Local build with a successful run.

Issue: #1