fix: Linked list was not setting tail correctly
Problem: First, the file handle was not being freed when the file is at
the start of the linked list. Second, the case when the head and tail
both should point to the same node was not covered. Second, each file
handle required a memory allocation on what could just be a fixed buffer.

Solution: Cover the two cases described above for `CloseHandle`, and
have a fixed buffer in `memfile` sized at `DVL_MAX_PATH`.

Testing: Testing: Built and verified that game starts and loads
correctly.

Issue: #3